Redirect to dash (#59)
